Refactor classes into packages. Add all classes to a package except for launcher and Duke.
