+++ src/main/java/duke/commands/TodoCommand.java
@@ -1,8 +1,8 @@
 package duke.commands;
 
-import duke.TaskList;
-import duke.Ui;
-import duke.Storage;
+import duke.logic.TaskList;
+import duke.logic.Ui;
+import duke.logic.Storage;
 import duke.exceptions.DukeEmptyInputException;
 import duke.tasks.Todo;
 import duke.exceptions.DukeInvalidInputException;
