+++ src/main/java/duke/commands/UnmarkCommand.java
@@ -2,9 +2,9 @@ package duke.commands;
 
 import duke.exceptions.DukeEmptyInputException;
 import duke.exceptions.DukeInvalidInputException;
-import duke.Storage;
-import duke.TaskList;
-import duke.Ui;
+import duke.logic.Storage;
+import duke.logic.TaskList;
+import duke.logic.Ui;
 
 /**
  * Represents the command to mark a task as not done.
