+++ src/main/java/duke/gui/Main.java
@@ -1,6 +1,8 @@
-package duke;
+package duke.gui;
 
 import java.io.IOException;
+
+import duke.Duke;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
