+++ src/main/java/duke/logic/Storage.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.logic;
 
 import java.io.File;
 import java.io.IOException;
