+++ src/main/java/duke/logic/TaskList.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.logic;
 
 import duke.exceptions.DukeEmptyInputException;
 import duke.exceptions.DukeException;
