+++ src/main/java/duke/logic/Ui.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.logic;
 
 import duke.exceptions.DukeException;
