+++ src/main/resources/view/MainWindow.fxml
@@ -10,7 +10,7 @@
 <!--Adapted from https://github.com/elizabethhky/ip/blob/master/src/main/resources/view/MainWindow.fxml-->
 <AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0"
             prefWidth="400.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1"
-            fx:controller="duke.MainWindow" style="-fx-background-color: #FFF8FF;">
+            fx:controller="duke.gui.MainWindow" style="-fx-background-color: #FFF8FF;">
     <children>
         <TextField fx:id="userInput" layoutY="558.0" onAction="#handleUserInput" prefHeight="41.0" prefWidth="324.0"
                    AnchorPane.bottomAnchor="1.0" AnchorPane.leftAnchor="1.0" AnchorPane.rightAnchor="70.0"
