+++ src/test/java/duke/ParserTest.java
@@ -9,6 +9,7 @@ import duke.commands.ListCommand;
 import duke.commands.MarkCommand;
 import duke.commands.UnmarkCommand;
 import duke.exceptions.DukeUnknownInputException;
+import duke.logic.Parser;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
